Fix duplicated code warnings
